job-service/requirements.txt
grpcio
grpcio-tools
protobuf
chromadb
sentence-transformers
PyPDF2
spacy
pydantic
python-docx
pdfminer.six
transformers
torch